requirements.txt
streamlit
google-generativeai
plotly
matplotlib
koreanize-matplotlib
python-dotenv
ollama
